refactor(chat): move ChatPage message-content helpers to their own module
ChatPage.tsx is a 10k-line page. The header menu, the row-key helpers
(msgIdentityKey, turnLeadKey, anchorAltIdFor, stableAnchorIdFor,
virtualKeyFor, uniqueRowKeys, messageRowKey), mintSendId, the knowledge
bubble chip and the user-bubble content renderers (renderUserContent and
its file/dir segment helpers) do not read page state. They move byte-for-
byte into pages/chat/ChatPageMessageContent.tsx; the page imports them
back. No behaviour, string, or wire format changes.

The page no longer re-exports the moved names. Every consumer of those
names is a test in this repository, so each one now imports the owning
module directly. The one remaining default export and the two pre-
existing re-exports (isBrowseCommand, PREFILL_STORAGE_KEY) are untouched.

First slice of the ChatPage split proposed in #7255, cut small enough
for one review pass and re-based on main's chat-core transport work.
