+++ website/src/test/ChatPageW3Coverage.test.tsx
@@ -247,7 +247,8 @@ Object.defineProperty(window, 'matchMedia', {
   })),
 })
 
-import ChatPage, { renderUserContent, virtualKeyFor, messageRowKey } from '../pages/ChatPage'
+import ChatPage from '../pages/ChatPage'
+import { renderUserContent, virtualKeyFor, messageRowKey } from '../pages/chat/ChatPageMessageContent'
 
 // --- Fixtures ---------------------------------------------------------------
