+++ website/src/test/SessionActionsMenu.popout.test.tsx
@@ -26,7 +26,7 @@ vi.mock('../api/client', () => ({
   },
 }))
 
-import { ChatHeaderMenu } from '../pages/ChatPage'
+import { ChatHeaderMenu } from '../pages/chat/ChatPageMessageContent'
 import { registerPopout, __resetForTests, __setNavigateForTests } from '../utils/chatPopout'
 import type { RootState } from '../store'
 import type { ChatSlot } from '../types'
